Update dotnet.yml: add test step, remove setup-dotnet
Updated the comment at the top of the dotnet.yml file to specify that the workflow builds a .NET project using MSBuild and added a link to the relevant GitHub documentation. Removed the setup for .NET using actions/setup-dotnet@v4 and specifying the .NET version 4.8.x in the jobs section. Added a Test step to the workflow to run msbuild /t:Test /p:Configuration=Release command for executing tests after building the project.