Add new workflow step and include application icon
Added a new step `ShowFiles` in `dotnet.yml` to list the contents of the `Fenix-Modbus\Fenix-Modbus` directory. Commented-out lines related to the `build` job were also added. In `FenixServer.csproj`, included an application icon (`HttpServer.ico`) by adding a new `<PropertyGroup>` with the `<ApplicationIcon>` tag. Additionally, added the `HttpServer.ico` file to the project as a content item within a new `<ItemGroup>`.