Update job parameters and modify directory listing steps
Updated the `name` to `Fenix-Modbus-Installer` and the `path` to `FenixSetup.msi`. Removed the `ShowFiles` step that listed the contents of the `Fenix-Modbus\Fenix-Modbus\` directory. Added a new `dir` command to list the contents of the same directory with a slight path format change. Added commented out lines for a `build` job on `windows-2022`.