Update job parameters and modify directory listing command
Updated the `name` and `path` parameters for the job to specify `Fenix-Modbus-Installer` and the path to `FenixSetup.msi` respectively. Modified the `ShowFiles` job to change the directory listing command from `dir \Fenix-Modbus\Fenix-Modbus\` to `dir D:\a\Fenix-Modbus\Fenix-Modbus\`. Added a commented-out section in the YAML file for a `build` job that runs on `windows-2022`.